Adding an empty impl for resource constraints for all platforms other than linux
